fix: レビュー指摘に対応（ReactNode import・frontmatter テスト・server-only）
