+++ lib/loadLegalDocument.test.ts
@@ -1,4 +1,7 @@
-import { describe, expect, it } from "vitest";
+import { describe, expect, it, vi } from "vitest";
+
+vi.mock("server-only", () => ({}));
+
 import { loadLegalDocument } from "./loadLegalDocument";
 import {
   GITHUB_DISCUSSIONS_URL,
