+++ lib/loadLegalDocument.ts
@@ -1,53 +1,17 @@
+import "server-only";
+
 import fs from "node:fs";
 import path from "node:path";
 import matter from "gray-matter";
+import { assertFrontmatter } from "@/lib/legalDocumentFrontmatter";
 import { substituteLegalDocumentPlaceholders } from "@/lib/legalDocumentPlaceholders";
+import type { LegalDocumentSlug, LoadedLegalDocument } from "@/lib/loadLegalDocument.types";
 
-/** 読み込み可能な法務ドキュメントの slug。 */
-export type LegalDocumentSlug = "terms" | "privacy";
-
-/** frontmatter から読み取る法務ページのメタ情報。 */
-export type LegalDocumentFrontmatter = {
-  title: string;
-  pageTitle: string;
-  description: string;
-  canonicalPath: string;
-  lastUpdated: string;
-};
-
-export type LoadedLegalDocument = LegalDocumentFrontmatter & {
-  content: string;
-};
-
-const REQUIRED_FRONTMATTER_KEYS: (keyof LegalDocumentFrontmatter)[] = [
-  "title",
-  "pageTitle",
-  "description",
-  "canonicalPath",
-  "lastUpdated",
-];
-
-function assertFrontmatter(
-  data: Record<string, unknown>,
-  slug: LegalDocumentSlug
-): LegalDocumentFrontmatter {
-  const missing = REQUIRED_FRONTMATTER_KEYS.filter(
-    (key) => typeof data[key] !== "string" || (data[key] as string).length === 0
-  );
-  if (missing.length > 0) {
-    throw new Error(
-      `content/legal/${slug}.md: missing or empty frontmatter: ${missing.join(", ")}`
-    );
-  }
-
-  return {
-    title: data.title as string,
-    pageTitle: data.pageTitle as string,
-    description: data.description as string,
-    canonicalPath: data.canonicalPath as string,
-    lastUpdated: data.lastUpdated as string,
-  };
-}
+export type {
+  LegalDocumentFrontmatter,
+  LegalDocumentSlug,
+  LoadedLegalDocument,
+} from "@/lib/loadLegalDocument.types";
 
 /**
  * `content/legal/{slug}.md` を読み込み、frontmatter と置換済み本文を返す。
