+++ package.json
@@ -31,6 +31,7 @@
     "react-markdown": "^10.1.0",
     "remark-gfm": "^4.0.1",
     "react-konva": "^19.2.3",
+    "server-only": "^0.0.1",
     "sonner": "^2.0.7",
     "tesseract.js": "^7.0.0",
     "zustand": "^5.0.12"
